+++ src/models/index.ts
@@ -2,27 +2,27 @@ import BigDipperNetwork from './big_dipper_network';
 // ================================
 // Transaction Message Types
 // ================================
-import MsgMultiSend from './msg/msg_multi_send';
-import MsgSend from './msg/msg_send';
-import MsgVerifyInvariant from './msg/msg_verify_invariant';
-import MsgFundCommunityPool from './msg/msg_fund_community_pool';
-import MsgSubmitProposal from './msg/msg_submit_proposal';
-import MsgSetWithdrawAddress from './msg/msg_set_withdrawal_address';
-import MsgWithdrawDelegatorReward from './msg/msg_withdrawal_delegaor_reward';
-import MsgCommunityPoolSpendProposal from './msg/msg_community_pool_spend_proposal';
-import MsgParameterChangeProposal from './msg/msg_parameter_change_proposal';
-import MsgSoftwareUpgradeProposal from './msg/msg_software_upgrade_proposal';
-import MsgTextProposal from './msg/msg_text_proposal';
-import MsgDeposit from './msg/msg_deposit';
-import MsgVote from './msg/msg_vote';
-import MsgUnjail from './msg/msg_unjail';
-import MsgCreateValidator from './msg/msg_create_validator';
-import MsgDelegate from './msg/msg_delegate';
-import MsgEditValidator from './msg/msg_edit_validator';
-import MsgRedelegate from './msg/msg_redelegate';
-import MsgUndelegate from './msg/msg_undelegate';
+import MsgMultiSend from './msg/bank/msg_multi_send';
+import MsgSend from './msg/bank/msg_send';
+import MsgVerifyInvariant from './msg/crisis/msg_verify_invariant';
+import MsgFundCommunityPool from './msg/distribution/msg_fund_community_pool';
+import MsgSubmitProposal from './msg/governance/msg_submit_proposal';
+import MsgSetWithdrawAddress from './msg/distribution/msg_set_withdrawal_address';
+import MsgWithdrawDelegatorReward from './msg/distribution/msg_withdrawal_delegaor_reward';
+import MsgCommunityPoolSpendProposal from './msg/governance/msg_community_pool_spend_proposal';
+import MsgParameterChangeProposal from './msg/governance/msg_parameter_change_proposal';
+import MsgSoftwareUpgradeProposal from './msg/governance/msg_software_upgrade_proposal';
+import MsgTextProposal from './msg/governance/msg_text_proposal';
+import MsgDeposit from './msg/governance/msg_deposit';
+import MsgVote from './msg/governance/msg_vote';
+import MsgUnjail from './msg/slashing/msg_unjail';
+import MsgCreateValidator from './msg/staking/msg_create_validator';
+import MsgDelegate from './msg/staking/msg_delegate';
+import MsgEditValidator from './msg/staking/msg_edit_validator';
+import MsgRedelegate from './msg/staking/msg_redelegate';
+import MsgUndelegate from './msg/staking/msg_undelegate';
 import MsgUnknown from './msg/msg_unknown';
-import MsgWithdrawValidatorCommission from './msg/msg_withdraw_validator_commission';
+import MsgWithdrawValidatorCommission from './msg/distribution/msg_withdraw_validator_commission';
 
 export {
   BigDipperNetwork,
