+++ src/models/msg/bank/msg_send.ts
@@ -1,7 +1,8 @@
 import numeral from 'numeral';
+import { Categories } from '../types';
 
 class MsgSend {
-  public category: 'bank' | 'crisis' | 'distribution' | 'governance' | 'slashing' | 'staking';
+  public category: Categories;
   public type: string;
   public fromAddress: string;
   public toAddress: string;
