+++ src/models/msg/distribution/msg_fund_community_pool.ts
@@ -1,27 +1,25 @@
 import numeral from 'numeral';
+import { Categories } from '../types';
 
-class MsgDeposit {
-  public category: 'bank' | 'crisis' | 'distribution' | 'governance' | 'slashing' | 'staking';
+class MsgFundCommunityPool {
+  public category: Categories;
   public type: string;
-  public proposalId: number | string;
   public depositor: string;
   public amount: {
     denom: string;
     amount: string | number;
-  }[];
+  }[]
 
   constructor(payload: any) {
-    this.category = 'governance';
+    this.category = 'distribution';
     this.type = payload.type;
-    this.proposalId = payload.proposalId;
     this.depositor = payload.depositor;
     this.amount = payload.amount;
   }
 
   static fromJson(json: any) {
-    return new MsgDeposit({
+    return new MsgFundCommunityPool({
       type: json['@type'],
-      proposalId: numeral(json.proposal_id).value(),
       depositor: json.depositor,
       amount: json?.amount.map((x) => {
         return ({
@@ -33,4 +31,4 @@ class MsgDeposit {
   }
 }
 
-export default MsgDeposit;
+export default MsgFundCommunityPool;
