+++ src/models/msg/governance/msg_vote.ts
@@ -1,7 +1,8 @@
 import numeral from 'numeral';
+import { Categories } from '../types';
 
 class MsgVote {
-  public category: 'bank' | 'crisis' | 'distribution' | 'governance' | 'slashing' | 'staking';
+  public category: Categories;
   public type: string;
   public proposalId: number | string;
   public voter: string;
