+++ src/models/msg/msg_unknown.ts
@@ -1,10 +1,12 @@
+import { Categories } from './types';
+
 class MsgUnknown {
-  public category: 'bank' | 'crisis' | 'distribution' | 'governance' | 'slashing' | 'staking' | 'unknown';
+  public category: Categories;
   public type: string;
   public data: JSON;
 
   constructor(payload: any) {
-    this.category = 'unknown';
+    this.category = 'others';
     this.type = payload.type;
     this.data = payload.data;
   }
