+++ src/models/msg/profiles/msg_save_profile.ts
@@ -1,0 +1,22 @@
+import { Categories } from '../types';
+
+class MsgUnjail {
+  public category: Categories;
+  public type: string;
+  public validatorAddress: string;
+
+  constructor(payload: any) {
+    this.category = 'profiles';
+    this.type = payload.type;
+    this.validatorAddress = payload.validatorAddress;
+  }
+
+  static fromJson(json: any) {
+    return new MsgUnjail({
+      type: json['@type'],
+      validatorAddress: json.validator_addr,
+    });
+  }
+}
+
+export default MsgUnjail;
