+++ src/models/msg/staking/msg_create_validator.ts
@@ -1,7 +1,8 @@
 import numeral from 'numeral';
+import { Categories } from '../types';
 
 class MsgCreateValidator {
-  public category: 'bank' | 'crisis' | 'distribution' | 'governance' | 'slashing' | 'staking';
+  public category: Categories;
   public type: string;
   public description: {
     moniker: string;
