+++ src/models/msg/staking/msg_edit_validator.ts
@@ -1,7 +1,8 @@
 import numeral from 'numeral';
+import { Categories } from '../types';
 
 class MsgEditValidator {
-  public category: 'bank' | 'crisis' | 'distribution' | 'governance' | 'slashing' | 'staking';
+  public category: Categories;
   public type: string;
   public description: {
     moniker: string;
