+++ src/models/msg/staking/msg_undelegate.ts
@@ -1,11 +1,11 @@
 import numeral from 'numeral';
+import { Categories } from '../types';
 
-class MsgRedelegate {
-  public category: 'bank' | 'crisis' | 'distribution' | 'governance' | 'slashing' | 'staking';
+class MsgUndelegate {
+  public category: Categories;
   public type: string;
   public delegatorAddress: string;
-  public validatorSrcAddress: string;
-  public validatorDstAddress: string;
+  public validatorAddress: string;
   public amount: {
     denom: string;
     amount: string | number;
@@ -15,17 +15,15 @@ class MsgRedelegate {
     this.category = 'staking';
     this.type = payload.type;
     this.delegatorAddress = payload.delegatorAddress;
-    this.validatorDstAddress = payload.validatorDstAddress;
-    this.validatorSrcAddress = payload.validatorSrcAddress;
+    this.validatorAddress = payload.validatorAddress;
     this.amount = payload.amount;
   }
 
   static fromJson(json: any) {
-    return new MsgRedelegate({
+    return new MsgUndelegate({
       type: json['@type'],
       delegatorAddress: json?.delegator_address,
-      validatorSrcAddress: json?.validator_src_address,
-      validatorDstAddress: json?.validator_dst_address,
+      validatorAddress: json?.validator_address,
       amount: {
         denom: json?.amount?.denom,
         amount: numeral(json?.amount?.amount).value(),
@@ -34,4 +32,4 @@ class MsgRedelegate {
   }
 }
 
-export default MsgRedelegate;
+export default MsgUndelegate;
