+++ src/models/msg/types.ts
@@ -1,0 +1,1 @@
+export type Categories = 'bank' | 'crisis' | 'distribution' | 'governance' | 'slashing' | 'staking' | 'profiles' | 'others';
